+++ example_async.py
@@ -22,6 +22,14 @@ async def run():
         # If the state was changed successfully, True is returned, otherwise False.
         print("success" if await dpa.device.async_set_led_setting(enable=True) else "failed")
 
+        # Check for new firmware versions
+        # {'result': 'UPDATE_NOT_AVAILABLE', 'new_firmware_version': ''}
+        print(await dpa.device.async_check_firmware_available())
+
+        # Start firmware update, if new version is available. Important: The response does not tell you anything about the
+        # success of the update itself.
+        print("update started" if await dpa.device.async_start_firmware_update() else "no update available")
+
         # Get details of wifi stations connected to the device: MAC address, wifi band and connection rates.
         # {'connected_stations':
         #   [