+++ tests/test_deviceapi.py
@@ -10,7 +10,7 @@ from devolo_plc_api.device_api.devolo_idl_proto_deviceapi_ledsettings_pb2 import
 from devolo_plc_api.device_api.devolo_idl_proto_deviceapi_wifinetwork_pb2 import (
     WifiConnectedStationsGet, WifiGuestAccessGet, WifiGuestAccessSetResponse, WifiNeighborAPsGet, WifiRepeatedAPsGet)
 from devolo_plc_api.exceptions.feature import FeatureNotSupported
-from devolo_plc_api.device_api.devolo_idl_proto_deviceapi_updatefirmware_pb2 import UpdateFirmwareCheck
+from devolo_plc_api.device_api.devolo_idl_proto_deviceapi_updatefirmware_pb2 import UpdateFirmwareCheck, UpdateFirmwareStart
 
 
 class TestDeviceApi:
@@ -123,6 +123,35 @@ class TestDeviceApi:
             assert firmware == MessageToDict(firmware_available,
                                              including_default_value_fields=True,
                                              preserving_proto_field_name=True)
+
+    @pytest.mark.asyncio
+    async def test_async_start_firmware_update(self, request):
+        firmware_update = UpdateFirmwareStart()
+
+        with patch("devolo_plc_api.clients.protobuf.Protobuf._async_get", new=CoroutineMock(return_value=Response)), \
+             patch("httpx.Response.aread", new=CoroutineMock(return_value=firmware_update.SerializeToString())):
+            device_api = DeviceApi(request.cls.ip,
+                                   AsyncClient(),
+                                   request.cls.device_info['_dvl-deviceapi._tcp.local.']['Path'],
+                                   request.cls.device_info['_dvl-deviceapi._tcp.local.']['Version'],
+                                   "update",
+                                   "password")
+
+            assert await device_api.async_start_firmware_update()
+
+    def test_start_firmware_update(self, request):
+        firmware_update = UpdateFirmwareStart()
+
+        with patch("devolo_plc_api.clients.protobuf.Protobuf._get", return_value=Response), \
+             patch("httpx.Response.read", return_value=firmware_update.SerializeToString()):
+            device_api = DeviceApi(request.cls.ip,
+                                   Client(),
+                                   request.cls.device_info['_dvl-deviceapi._tcp.local.']['Path'],
+                                   request.cls.device_info['_dvl-deviceapi._tcp.local.']['Version'],
+                                   "update",
+                                   "password")
+
+            assert device_api.start_firmware_update()
 
     @pytest.mark.asyncio
     async def test_async_get_wifi_connected_station(self, request):