+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 #### Device API
 
 - Check for firmware updates
+- Start firmware updates
 
 ### Fixed